+++ blogging-platform/src/pages/Register.tsx
@@ -46,6 +46,7 @@ const Register: React.FC = () => {
                         type='text'
                         value={username}
                         onChange={(e) => setUsername(e.target.value)}
+                        maxLength={30}
                         required
                     />
                 </div>
@@ -55,6 +56,7 @@ const Register: React.FC = () => {
                         type='password'
                         value={password}
                         onChange={(e) => setPassword(e.target.value)}
+                        maxLength={30}
                         required
                     />
                 </div>